backend/requirements.txt
# =============================================================================
# MindMesh Backend Dependencies
# =============================================================================

# Core FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy[asyncio]==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Vector Database
pgvector==0.2.4

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# AI/ML Dependencies
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.1.1
langchain-community==0.0.10
langgraph==0.0.20
langsmith==0.0.69

# Vector Embeddings
sentence-transformers==2.2.2
openai==1.6.1
anthropic==0.8.1

# Storage
boto3==1.34.0
minio==7.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0
python-dotenv==1.0.0

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
